+++ changes/changelog.md
@@ -2,6 +2,16 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.3.0] - 2019-06-18
+### Added
+- `!food` to build and check a food order.
+To use this functionality do:
+* `!food <food>` to order `<food>`. Example: `!food Mushroom Toastie`
+* `!food` to check the order
+* `!food done` to clear the order
+* `!food cancel` to cancel anything that you've ordered
+
+
 ## [1.3.0] - 2019-06-12
 ### Added
 - `!ignore @user` to make Henchman ignore a user.